move scraper
